Fix typo: "as much slots" -> "as many slots"
